backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
google-cloud-bigquery==3.14.1
google-cloud-aiplatform==1.38.1
redis==5.0.1
pandas==2.1.4
numpy==1.26.3
hypothesis==6.92.1
pytest==7.4.3
pytest-asyncio==0.23.3
httpx==0.26.0
python-dotenv==1.0.0
